% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptual_work_adj.R
\name{conceptual_work_adj}
\alias{conceptual_work_adj}
\title{Measure normalized conceptual work}
\usage{
conceptual_work_adj(freq, cos_dist, threshold)
}
\description{
Measure normalized conceptual work
}
